+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionUtils.kt
@@ -813,13 +813,11 @@ fun ActionData.canBeHeldDown(): Boolean = when (this) {
 fun ActionData.canUseImeToPerform(): Boolean = when (this) {
     is ActionData.InputKeyEvent -> !useShell
     is ActionData.Text -> true
-    is ActionData.MoveCursorToEnd -> true
     else -> false
 }
 
 fun ActionData.canUseShizukuToPerform(): Boolean = when (this) {
     is ActionData.InputKeyEvent -> true
-    is ActionData.MoveCursorToEnd -> true
     else -> false
 }