+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - #1711 major refactoring of the entire codebase into separate Gradle modules.
 - #1701 improve the order of the actions and categories
+- Key Mapper keyboard or Shizuku are no longer required for the action to move the cursor to the end
 
 ## Bug fixes